pyproject.toml
[project]
name = "tasks-tw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ar3ac", email = "luca.marrazzo@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth>=2.45.0",
    "google-auth-oauthlib>=1.2.2",
]

[project.scripts]
tasks-tw = "tasks_tw:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
